Added Util.ModuleTools
