+++ PynPoint/IOmodules/Hdf5Reading.py
@@ -9,7 +9,7 @@ import h5py
 import numpy as np
 
 from PynPoint.Core.Processing import ReadingModule
-from PynPoint.Util.Progress import progress
+from PynPoint.Util.ModuleTools import progress
 
 
 class Hdf5ReadingModule(ReadingModule):
