+++ PynPoint/ProcessingModules/PSFpreparation.py
@@ -12,9 +12,9 @@ import numpy as np
 
 from scipy import ndimage
 
-from PynPoint.Util.Progress import progress
 from PynPoint.Core.Processing import ProcessingModule
 from PynPoint.ProcessingModules.ImageResizing import CropImagesModule, ScaleImagesModule
+from PynPoint.Util.ModuleTools import progress, memory_frames
 
 
 class PSFpreparationModule(ProcessingModule):
@@ -348,18 +348,7 @@ class SortParangModule(ProcessingModule):
 
         nimages = self.m_image_in_port.get_shape()[0]
 
-        if memory == 0 or memory >= nimages:
-            frames = [0, nimages]
-
-        else:
-            frames = np.linspace(0,
-                                 nimages-nimages%memory,
-                                 int(float(nimages)/float(memory))+1,
-                                 endpoint=True,
-                                 dtype=np.int)
-
-            if nimages%memory > 0:
-                frames = np.append(frames, nimages)
+        frames = memory_frames(memory, nimages)
 
         for i, _ in enumerate(frames[:-1]):
             progress(i, len(frames[:-1]), "Running SortParangModule...")
