--- PynPoint/Util/Progress.py
@@ -1,8 +1,0 @@
-import sys
-
-def progress(current, total, message):
-    fraction = float(current)/float(total)
-    percentage = round(fraction*100., 1)
-
-    sys.stdout.write("%s %s%s \r" % (message, percentage, "%"))
-    sys.stdout.flush()
